backend/requirements.txt
flask
flask_sqlalchemy
flask_migrate
pymysql
python-dotenv
pymongo
